refactor: split plugin apply to configurators
